Adventures: Dump + Summary
Tweak the Rakefile so that when an adventure is compiled without DBG or
PACK options it will generate a complete dump file (.lis) along with a
summary file (.sum) by invoking the ALAN compiler with:

    -listing -height 0 -dump ! -summary <fname>.alan

These option will automatically generate a `<fname>.lis` listing file
with the dump info, and print to STDOUT the summary info, which is
captured by the Rake task and written to a `<fname>.sum` file.

These summary and dump files (ignored by Git) provide useful information
to aid inspection of .a3c storyfiles.